// src/Web/EduArkApp/src/app/core/models/class/class.master.data.model.ts
import { DropDownModel } from './../common/drop.down.model';
export class ClassMasterDataModel {
	currentAcademicYear: number;
	academicYears: DropDownModel[];
	academicLevels: DropDownModel[];
	classNames: DropDownModel[];
	classCategories: DropDownModel[];
	languageStreams: DropDownModel[];
	allTeachers: DropDownModel[];
	examTypes: DropDownModel[];
}